Create a QuickFormTestBase class that other quick form tests can extend from #650
